Doxgyen: added docs. cleaned up an error

+++ Electrode/src/Electrode_Integrator.cpp
@@ -2340,7 +2340,7 @@ void Electrode_Integrator::printElectrode(int pSrc, bool subTimeStep)
         printf("                ChemModel Type = %3d , DomainNumber = %2d , CellNumber = %2d , IntegrationCounter = %d\n",
                electrodeChemistryModelType_, electrodeDomainNumber_, electrodeCellNumber_, counterNumberIntegrations_);
 	printf("                numIntegrationSubCycles = %d, SubIntegrationCounter = %d\n",
-	       numIntegrationSubCycles_final_final_, counterNumberSubIntegrations_);
+	       static_cast<int>(numIntegrationSubCycles_final_final_), counterNumberSubIntegrations_);
     }
     printf("   ==============================================================================================\n");
     printf("\n");
